Refactor code Change some styles Fix some bugs
