+++ src/components/Header/Header.module.scss
@@ -1,3 +1,5 @@
+@import '../../styles/vars.scss';
+
 .header {
   display: flex;
   flex: 0 1 95px;
@@ -5,17 +7,21 @@
   justify-content: space-between;
   padding: 0px 100px;
   color: #fff;
-  box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.15);
+  box-shadow: 0px 4px 10px rgba(black, 0.15);
 }
 
 .logo {
-  color: #23335d;
+  color: $flamingo;
   font-weight: bold;
   font-size: 1.5rem;
   line-height: 28px;
 
   .logoIcon {
     margin-right: 10px;
+
+    path {
+      fill: $flamingo;
+    }
   }
 }
 
@@ -23,21 +29,16 @@
   display: flex;
   align-items: center;
   padding: 10px;
-  border: 1px solid #23335d;
+  border: 1px solid $lightgrey;
   border-radius: 3px;
-  color: #23335d;
+  color: $shuttleGrey;
   font-weight: bold;
   font-size: 12px;
   line-height: 14px;
   cursor: pointer;
   transition: background 0.2s;
 
-  &:hover {
-    background: #23335d;
-    color: #fff;
-  }
-
-  .favesBtnIcon {
+  .starIcon {
     margin-right: 4px;
   }
 }
