+++ src/components/Search/Search.module.scss
@@ -21,13 +21,8 @@
     width: 100%;
     height: 100%;
 
-    // placing placeholder in the top is devided between css and js, css for focus and js for typing
     &:focus + .placeholder {
-      top: 2px;
-      color: rgba($fiord, 0.7);
-      font-weight: normal;
-      font-size: 12px;
-      line-height: 14px;
+      display: none;
     }
   }
 }
@@ -43,7 +38,7 @@
   }
 
   path {
-    fill: $location-icon-color;
+    fill: $lightgrey;
   }
 }
 
@@ -66,19 +61,12 @@
 .placeholder {
   position: absolute;
   z-index: 1;
+  display: block;
   margin-left: 22px;
   color: rgba($arsenic, 0.8);
-  font-weight: bold;
+  font-weight: 500;
   font-size: 14px;
   line-height: 16px;
-
-  &Top {
-    top: 2px;
-    color: rgba($fiord, 0.7);
-    font-weight: normal;
-    font-size: 12px;
-    line-height: 14px;
-  }
 }
 
 .suggestions {
@@ -97,13 +85,10 @@
   cursor: pointer;
 
   &:hover {
-    color: black;
-    font-weight: bold;
-    text-decoration: underline;
-  }
-
-  .toHightlight {
-    color: red;
+    font-weight: 600;
+    font-style: normal;
+    font-size: 14px;
+    line-height: 16px;
   }
 }
