+++ src/components/index.jsx
@@ -1,0 +1,2 @@
+export * from './Header/Header';
+export * from './Search/Search';
