+++ src/containers/Home/Home.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Search from '../../components/Search/Search';
+import { Search } from '../../components';
 import styles from './Home.module.scss';
 
 export const Home = () => {
