+++ src/containers/Home/Home.module.scss
@@ -7,11 +7,10 @@
 }
 
 .pageContent {
-  flex: 1 1;
+  flex: 0 1 502px;
   padding: 100px;
   border-radius: 30px;
-  background: url('../../assets/background.jpg') center;
-  background-size: 100% 100%;
+  background: url('../../assets/background.png') center;
   background-repeat: no-repeat;
   color: white;
