+++ src/styles/vars.scss
@@ -1,7 +1,8 @@
 $fiord: #4a5063;
 $regal-blue: #23335d;
-$location-icon-color: #dbdfe1;
+$lightgrey: #dbdfe1;
 $flamingo: #e6605b;
 $arsenic: #384047;
 $sapphire: #092c4c;
 $zircon: #dbdfe1;
+$shuttleGrey: #60666c;
